MV: generateSessionId to separate file
